+++ fw/peripherals/hal/spi_helper.h
@@ -8,9 +8,11 @@
 #ifndef FW_PERIPHERALS_HAL_SPI_HELPER_H_
 #define FW_PERIPHERALS_HAL_SPI_HELPER_H_
 
+#include <stdint.h>
+#include <stddef.h>
 
 void spi_helper_init_handle();
 
-uint8_t spi_helper_transfer(uint8_t nBytes, uint8_t* txBufferPointer, uint8_t* rxBufferPointer, uint8_t CS_pin);
+uint8_t spi_helper_transfer(size_t nBytes, uint8_t* txBufferPointer, uint8_t* rxBufferPointer, uint8_t CS_pin);
 
 #endif /* FW_PERIPHERALS_HAL_SPI_HELPER_H_ */
